activePictureId and apiState are in drawingStateStore
Moving these states inside of drawingStateStore simplifies the store
architecture because these properties are not data (as should be stored
in PictureStore), but are instead states of the App. This also continues
to make the dependency clean where the state of the App is always
dependent on data store and not the other way around.